+++ packages/scraper-python/src/mcpscraper/__init__.py
@@ -2,8 +2,17 @@ from .client import ScraperClient
 from .errors import ScraperApiError
 from .mcp_client import McpToolsClient, McpToolError
 from ._mcp_generated_client import MCP_TOOL_BINDINGS, MCP_TOOL_COUNT
+from .assistant_types import (
+    AssistantApprovalDecision,
+    AssistantCommand,
+    AssistantCursor,
+    AssistantIdempotencyKey,
+    AssistantPageRequest,
+)
 
 __all__ = [
     "ScraperClient", "ScraperApiError", "McpToolsClient", "McpToolError",
     "MCP_TOOL_BINDINGS", "MCP_TOOL_COUNT",
+    "AssistantApprovalDecision", "AssistantCommand", "AssistantCursor",
+    "AssistantIdempotencyKey", "AssistantPageRequest",
 ]